Refactor article and build code
